+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,221 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.10.4...@synapsecns/synapse-interface@0.10.5) (2024-03-20)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.10.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.10.3...@synapsecns/synapse-interface@0.10.4) (2024-03-20)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.10.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.10.2...@synapsecns/synapse-interface@0.10.3) (2024-03-20)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.10.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.10.1...@synapsecns/synapse-interface@0.10.2) (2024-03-19)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.10.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.10.0...@synapsecns/synapse-interface@0.10.1) (2024-03-19)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+# [0.10.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.10...@synapsecns/synapse-interface@0.10.0) (2024-03-18)
+
+
+### Features
+
+* **synapse-interface:** create template event maintenance components, remove ecotone + metis event maintenance components ([#2294](https://github.com/synapsecns/sanguine/issues/2294)) ([7780b01](https://github.com/synapsecns/sanguine/commit/7780b010bd5552b2e53db9bf4814dc80318cb57c))
+
+
+
+
+
+## [0.9.10](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.9...@synapsecns/synapse-interface@0.9.10) (2024-03-18)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.9.9](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.8...@synapsecns/synapse-interface@0.9.9) (2024-03-16)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.9.8](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.7...@synapsecns/synapse-interface@0.9.8) (2024-03-14)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.9.7](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.6...@synapsecns/synapse-interface@0.9.7) (2024-03-14)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.9.6](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.5...@synapsecns/synapse-interface@0.9.6) (2024-03-14)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.9.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.4...@synapsecns/synapse-interface@0.9.5) (2024-03-14)
+
+
+### Bug Fixes
+
+* **synapse-interface:** fixes explorer images ([#2292](https://github.com/synapsecns/sanguine/issues/2292)) ([98f48b1](https://github.com/synapsecns/sanguine/commit/98f48b165f45b2a9b8f291e3020872029ae20abb))
+
+
+
+
+
+## [0.9.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.3...@synapsecns/synapse-interface@0.9.4) (2024-03-14)
+
+
+### Bug Fixes
+
+* **synapse-interface:** deprecate unused dependencies ([#2287](https://github.com/synapsecns/sanguine/issues/2287)) ([e14071b](https://github.com/synapsecns/sanguine/commit/e14071bf1014050036b5e2e0858aa1e84aad2e18))
+
+
+
+
+
+## [0.9.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.2...@synapsecns/synapse-interface@0.9.3) (2024-03-14)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.9.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.1...@synapsecns/synapse-interface@0.9.2) (2024-03-14)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.9.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.9.0...@synapsecns/synapse-interface@0.9.1) (2024-03-14)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+
+# [0.9.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.8.0...@synapsecns/synapse-interface@0.9.0) (2024-03-13)
+
+
+### Features
+
+* **synapse-interface:** use both rebate displays in bridge card ([#2276](https://github.com/synapsecns/sanguine/issues/2276)) ([45fbfdc](https://github.com/synapsecns/sanguine/commit/45fbfdc245647b8db87b889e252fa4aece49f501))
+
+
+
+
+
+# [0.8.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.7.1...@synapsecns/synapse-interface@0.8.0) (2024-03-13)
+
+
+### Features
+
+* **synapse-interface:** ecotone and metis upgrade downtime, remove eth dencun ([#2274](https://github.com/synapsecns/sanguine/issues/2274)) ([f527ce5](https://github.com/synapsecns/sanguine/commit/f527ce53c883672232f43da90f0614a8a580442a))
+
+
+
+
+
+## [0.7.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.7.0...@synapsecns/synapse-interface@0.7.1) (2024-03-13)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+
+# [0.7.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.6.2...@synapsecns/synapse-interface@0.7.0) (2024-03-13)
+
+
+### Features
+
+* **synapse-interface:** maintenance notification components + eth dencun pause ([#2260](https://github.com/synapsecns/sanguine/issues/2260)) ([9ffae5b](https://github.com/synapsecns/sanguine/commit/9ffae5b640f904818fdd4b198c069042f05ffb29))
+
+
+
+
+
+## [0.6.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.6.1...@synapsecns/synapse-interface@0.6.2) (2024-03-13)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.6.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.6.0...@synapsecns/synapse-interface@0.6.1) (2024-03-13)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+# [0.6.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.5.1...@synapsecns/synapse-interface@0.6.0) (2024-03-12)
+
+
+### Features
+
+* **synapse-interface:** complete countdown ([#2254](https://github.com/synapsecns/sanguine/issues/2254)) ([fcb685f](https://github.com/synapsecns/sanguine/commit/fcb685f12f70f3623e16bb0260503156568b3f83))
+
+
+
+
+
+## [0.5.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.5.0...@synapsecns/synapse-interface@0.5.1) (2024-03-12)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 # [0.5.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.4.0...@synapsecns/synapse-interface@0.5.0) (2024-03-08)